@@ -8,7 +8,7 @@ module.exports = {
     process: true,
   },
   extends: ["eslint:recommended", "plugin:react/recommended"],
-  parser: "babel-eslint",
+  parser: "@babel/eslint-parser",
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
